Adds basic editor support to defaultLayout Moves applying default layout logic into layout.ts
